Converts transform package to module Renames package to _transform, adds transform module, removes redundant init, and cleans up imports.
